requirements.txt
appdirs==1.4.3
APScheduler==3.3.1
cffi==1.10.0
dj-static==0.0.6
Django==1.11
django-extensions==1.7.8
django-solo==1.1.2
greenlet==0.4.12
gunicorn==19.7.1
packaging==16.8
psycopg2==2.7.1
pycparser==2.17
pyparsing==2.2.0
pytz==2017.2
readline==6.2.4.1
six==1.10.0
static3==0.7.0
tzlocal==1.4
virtualenv==15.1.0
